Refactoring of some commands
